split get_exchanges_and_currencies() into smaller functions
This functions has different behaviour depending if it is run for the first time or not.
Exctract some code into smaller function. One function does the querying and returns a dict.
Another function writes this to a file.

Test Plan:
delete the currencies.json file, laucnh the program and watch it be recreated