enable fiat conversion again
Show the widgets again in the settings menu.
For new users migrating from Electron Cash, do not disable it when copying the user directory.

For this commit, the APIs are still set to query BCH

Test Plan:
Run the program from the repository, check that the fiat options work again.