use the default directory when creating new wallets
The previous behavior was to create new wallets in the same directory as the current wallet, even if this wallet was a wallet open in an arbitrary external path.

This changes the behavior to use either the wallet path explicitely specified on the command line (-w or --wallet argument) or the default wallet directory in the user's data directory.

Remove also a piece of code providing compatiblity with an ancient version of electrum (1.9 released in 2013).

Closes #10

Test Plan:
Run `electrum-abc`, open a wallet file in an external directory unrelated to electrum-abc (for instance in the Electron Cash data directory), then create a new wallet and verify that it is created in the normal .electrum-abc directory.