change the description for the payment when donation to Fulcrum
When a Fulcrum server does not specify a donation address in its
configuration file, the server broadcasts the donation address
for Fulcrum development. Make that clear in the description field
when a user clicks on "Donate to server" in the menu.

Test plan:
Connect to electrum.bitcoinabc.org, click on "donate to server",
check the description field.