use a much simpler way to set the default command as "gui"
No need to monkey patch a third party library, risking to break it when they update.
Just set the default command explicitly if none is provided.

Test plan:
run the program from the command line without specifying any argument, verify that it is the same behavior as running `electrum-abc gui`.

Search for the 'cmd' string in all the modules and make sure no other module needs it, beside the main script.