reduce repetition of password prompts
Change the default `prompt_password` arguments to reflect the majority of uses.
Create an alias for the specific case of prompting a password for creating a new wallet or restoring a wallet from seed. In this case the password has a specific purpose (encrypting a wallet file) and requires confirmation.