fix get_tx_status
I broke the get_tx_status method when I removed an unused status from a list
in 1538289f386ddd56836b58bd43c77623e2115c13.

This fixes it by removing that list, which seems to be an unnecessary indirection,
and using the opportunity to documentate the function.

Test  plan:
send 2 transactions in a row, the second one by spending the output of the first.
Wait for 6 confirmations, so that these transaction cover all the possible
status values.