Patch error when trying to connect to satochip
During wallet creation, the following error sometimes pops to the user:
"Failed to create a client for this device. Make sure it is in a correct state".

This error does not always show, and even when it does, it does not stop the wallet creation.

This issue is generated in the method 'has_usable_connection_with_device()'.
The reason is that sometimes the card is not ready when the wizard tries to connect to it.
Retrying after a short delay is enough to solve the issue.