Windows: Use PyInstaller version 4.2
PyInstaller 3.6 fixed the MinGW build issue we used a patch for:

https://github.com/pyinstaller/pyinstaller/commit/6df55e45876968537585c197f2b86f67d0242305

but version 4.2 broke the MinGW build again:

https://github.com/pyinstaller/pyinstaller/commit/99db9ab98d63b028724d38f241424d21df44afcf

this fails with `windres: no resources` because the preprocessed RC file
is essentially empty. Windows RC seems to ignore this. To fix it we add
a patch that just removes the RC file build:

https://github.com/EchterAgo/pyinstaller/commit/d6f3d02365ba68ffc84169c56c292701f346110e

This is a backport of https://github.com/Electron-Cash/Electron-Cash/pull/2161