add a context menu action to dump utxo details in the Coins tab
The coin informations are dumped to a json file specifid by the user.

This commit also fixes a code smell: in  `create_menu`, line 229,
`coins` was a filter object (iterator) with a truth value of True no
matter is there were any coins or not. Now it is converted to a list,
which does not impact the performance as is has to be generated anyway
on line 233 to filter spendable_coins.

Test plan:

Select a single coin and use the menu action.
Select multiple coins and use the menu action. Include frozen coins, slp
tokens, regular coins ...

Inspect the output files.