implement a "build avalanche proof" action in the Coins dialog
The user selects one or more coins to use them as stakes for building an
avalanche proof, then right clicks and chooses the action in the
dropdown menu. A dialog opens with fields to fill: master key,
proof expiration date, sequence number.
Clicking the generate button causes the proof to be generated, if the
master key is a valid WIF private key or a hexadecimal public key.