export utxo details: don't change the field names
Previously I changed some field names when exporting UTXOs to use the
same vocabulary as `bitcoin-cli buildavalancheproof`, but this makes it
harder to use the same tool for building the proof from a selection of
coins from the coins tab or from a the exported json file.

Use the Electrum vocabulary (txid: prevout_hash...). Only the address
needs to be serialized (and deserialized again on the other side of the
pipe)