remove android and ios codebase
There is no plan to release mobile apps for Electrum ABC at the moment,
and these codebases are falling behind on backports anyway.

If we ever plan to support ios and android, it will be best to copy the
latest versions of these directories and then start patching them for
BCHA.

Remove also a tox related file. Using tox for testing has been dropped a
while ago.