Install Wizard: Add derivation path scanner (#2199)
* Install Wizard: Add derivation path scanner

This allows a user to scan a set of known derivation paths for
transactions when restoring a wallet from a seed.

* Use global network instance

* Do not touch GUI stuff from other thread

The callback will not emit a signal and the table update will be done in
the main thread.

Also add early return if there isn't any network so we don't have to
mess around with any keys, wallets or stuff like that if we can't really
scan.