fix thousands grouping for locale unaware systems
Some users report that the amounts are not formatted with thousands
grouping. It seems that setting the locale does not always works.

This diff improves the situation in the following way:
 - change how the locale is set, by using the empty-string in `locale.setlocale` (from the doc: An empty string specifies the user’s default settings.)
 - test whether the locale settings actually work for thousands grouping in the `format_satoshis`
 - if the locale settings don't work as expected, fall back to python's local-unaware way of printing amounts with numbers grouping (using a "," in the format string).

Adapt tests so that they work for both situations (`LOCALE_HAS_THOUSANDS_GROUPING` being `True` or `False`).

Note that the `False` case is tested for the C locale.