change threshold for long address in "transaction details"
We recently started showing the prefix in all CashAddresses everywhere in the GUI.
As a result, the "tx details" dialog started splitting outputs on 2 lines.

A cash address with a "ecash:" prefix has a length of 48.
42 is the length of an address with prefix omitted.

This change puts outputs back on a single line with the "ecash:"
cashaddress format.